Fixed circular dependency, broken test.
